requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.7.0=pyhe01879c_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
attr=2.5.1=h166bdaf_1
brotli-python=1.1.0=py312h2ec8cdc_2
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.4.26=pyhd8ed1ab_0
cffi=1.17.1=py312h06ac9bb_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
click=8.2.1=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
colorcet=3.1.0=pypi_0
contourpy=1.3.2=pypi_0
cpython=3.12.10=py312hd8ed1ab_0
cuda-crt-tools=12.9.41=ha770c72_0
cuda-cudart=12.9.37=h5888daf_0
cuda-cudart_linux-64=12.9.37=h3f2d84a_0
cuda-cuobjdump=12.9.26=hbd13f7d_0
cuda-cupti=12.9.19=h9ab20c4_0
cuda-nvcc-tools=12.9.41=he02047a_0
cuda-nvdisasm=12.9.19=hbd13f7d_0
cuda-nvrtc=12.9.41=h5888daf_0
cuda-nvtx=12.9.19=h5888daf_0
cuda-nvvm-tools=12.9.41=he02047a_0
cuda-version=12.9=h4f385c5_3
cudnn=9.10.1.4=h7646684_0
cusparselt=0.7.0.0=hcd2ec93_0
cycler=0.12.1=pypi_0
diffusers=0.33.1=pypi_0
docker-pycreds=0.4.0=pypi_0
drjit=1.0.5=pypi_0
einops=0.8.1=pypi_0
filelock=3.18.0=pyhd8ed1ab_0
fonttools=4.58.0=pypi_0
fsspec=2025.5.0=pyhd8ed1ab_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gmp=6.3.0=hac33072_2
gmpy2=2.2.1=py312h7201bc8_0
h2=4.2.0=pyhd8ed1ab_0
h5py=3.13.0=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
huggingface_hub=0.31.4=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=75.1=he02047a_0
idna=3.10=pyhd8ed1ab_1
imageio=2.37.0=pypi_0
importlib-metadata=8.7.0=pypi_0
jinja2=3.1.6=pyhd8ed1ab_0
kiwisolver=1.4.8=pypi_0
ld_impl_linux-64=2.43=h712a8e2_4
libabseil=20240722.0=cxx17_hbbce691_4
libblas=3.9.0=31_h59b9bed_openblas
libcap=2.75=h39aace5_0
libcblas=3.9.0=31_he106b2a_openblas
libcublas=12.9.0.13=h9ab20c4_0
libcudnn=9.10.1.4=h4840ae0_0
libcudnn-dev=9.10.1.4=hcd2ec93_0
libcudss0=0.4.0.2=he55f5cd_2
libcufft=11.4.0.6=h5888daf_0
libcufile=1.14.0.30=h628e99a_0
libcurand=10.3.10.19=h9ab20c4_0
libcusolver=11.7.4.40=h9ab20c4_0
libcusparse=12.5.9.5=h5888daf_0
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgcrypt-lib=1.11.1=hb9d3cd8_0
libgfortran=15.1.0=h69a702a_2
libgfortran5=15.1.0=hcea5267_2
libgomp=15.1.0=h767d61c_2
libgpg-error=1.55=h3f2d84a_0
libiconv=1.18=h4ce23a2_1
liblapack=3.9.0=31_h7ac8fdf_openblas
libllvm20=20.1.5=he9d0ab4_0
liblzma=5.8.1=hb9d3cd8_1
libmagma=2.8.0=h566cb83_2
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libnvjitlink=12.9.41=h5888daf_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libprotobuf=5.28.3=h6128344_1
libsqlite=3.49.2=hee588c1_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libsystemd0=257.4=h4e0b6ca_1
libtorch=2.6.0=cuda126_generic_h4a15719_200
libudev1=257.4=hbe16f8c_1
libuuid=2.38.1=h0b41bf4_0
libuv=1.50.0=hb9d3cd8_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.13.8=h4bc477f_0
libzlib=1.3.1=hb9d3cd8_2
lz4-c=1.10.0=h5888daf_1
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py312h178313f_1
matplotlib=3.10.3=pypi_0
mdurl=0.1.2=pypi_0
mitsuba=3.6.4=pypi_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
nccl=2.26.6.1=ha44e49d_0
ncurses=6.5=h2d0b736_3
networkx=3.4.2=pyh267e887_2
nibabel=5.3.2=pypi_0
nomkl=1.0=h5ca1d4c_0
numpy=2.2.6=py312h72c5963_0
omegaconf=2.3.0=pypi_0
openexr=3.3.3=pypi_0
openssl=3.5.0=h7b32b05_1
optree=0.15.0=py312h68727a3_0
packaging=25.0=pyh29332c3_1
pillow=11.2.1=pypi_0
pip=25.1.1=pyh8b19718_0
platformdirs=4.3.8=pypi_0
protobuf=6.31.0=pypi_0
psutil=7.0.0=py312h66e93f0_0
pybind11=2.13.6=pyh1ec8472_2
pybind11-global=2.13.6=pyh415d2e4_2
pycparser=2.22=pyh29332c3_1
pydantic=2.11.4=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pyha55dd90_7
python=3.12.10=h9e4cc4f_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.12=7_cp312
pytorch=2.6.0=cuda126_generic_py312_h460ed29_200
pyyaml=6.0.2=py312h178313f_2
rdma-core=57.0=h5888daf_0
readline=8.2=h8c095d6_2
regex=2024.11.6=pypi_0
requests=2.32.3=pyhd8ed1ab_1
rich=14.0.0=pypi_0
safetensors=0.5.3=py312h12e396e_0
scipy=1.15.3=pypi_0
sentry-sdk=2.29.1=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=80.8.0=pyhff2d567_0
six=1.17.0=pypi_0
sleef=3.8=h1b44611_0
smmap=5.0.2=pypi_0
sympy=1.14.0=pyh2585a3b_105
tk=8.6.13=noxft_h4845f30_101
tokenizers=0.21.1=pypi_0
tqdm=4.67.1=pyhd8ed1ab_1
transformers=4.52.2=pypi_0
triton=3.2.0=cuda126py312h5a3d8a8_1
typing-extensions=4.13.2=h0e9735f_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025b=h78e105d_0
urllib3=2.4.0=pyhd8ed1ab_0
wandb=0.19.11=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
yaml=0.2.5=h7f98852_2
zipp=3.21.0=pypi_0
zstandard=0.23.0=py312h66e93f0_2
zstd=1.5.7=hb8e6e7a_2
